+++ lib/lint_hard.dart
@@ -1,0 +1,1 @@
+export 'src/lint_hard_plugin.dart';